+++ lib/common/styles.dart
@@ -11,3 +11,8 @@ const double kAppBarMidHeight = 256.0;
 const double personDelegateIconHeight = 40.0;
 
 const double homeScreenBottomBarHeight = 50.0;
+class Constants {
+  Constants._();
+  static const double padding = 20;
+  static const double avatarRadius = 80;
+}
