+++ lib/ui/home/chats_tab.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
 import 'package:provider/provider.dart';
 import 'package:retroshare/provider/subscribed.dart';
 import 'package:retroshare/services/chat.dart';
@@ -16,8 +15,7 @@ class ChatsTab extends StatelessWidget {
     return SafeArea(
       top: false,
       bottom: false,
-      child: Stack(
-        children: <Widget>[
+      child:
           Consumer<ChatLobby>(
             builder: (context, chatsList, _) {
               if (chatsList.subscribedlist != null &&
@@ -65,38 +63,30 @@ class ChatsTab extends StatelessWidget {
                   ],
                 );
              
-                return Container();
-            },
-          ),
-          Consumer<ChatLobby>(
-            builder: (context, chatsList, _) {
-              if (chatsList.subscribedlist == null ||
-                  chatsList.subscribedlist?.isEmpty)
                 return Center(
-                  child: SizedBox(
-                    width: 200,
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: <Widget>[
-                        Image.asset('assets/icons8/pluto-sign-in.png'),
-                        Padding(
-                          padding: EdgeInsets.symmetric(vertical: 25),
-                          child: Text(
-                            "Looks like there aren't any subscribed chats",
-                            style: Theme.of(context).textTheme.body2,
-                            textAlign: TextAlign.center,
-                          ),
+                child: SizedBox(
+                  width: 200,
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    children: <Widget>[
+                      Image.asset('assets/icons8/pluto-sign-in.png'),
+                      Padding(
+                        padding: EdgeInsets.symmetric(vertical: 25),
+                        child: Text(
+                          "Looks like there aren't any subscribed chats",
+                          style: Theme.of(context).textTheme.body2,
+                          textAlign: TextAlign.center,
                         ),
-                      ],
-                    ),
+                      ),
+                    ],
                   ),
-                );
-              else
-                return Container();
+                ),
+              );
             },
           ),
-        ],
-      ),
+          
+        
+      
     );
   }
 }
