+++ lib/ui/room/room_screen.dart
@@ -1,11 +1,11 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:retroshare/Middleware/register_chat_event.dart';
 import 'package:retroshare/common/styles.dart';
 import 'package:retroshare/model/cache.dart';
 import 'package:retroshare/provider/FriendsIdentity.dart';
 import 'package:retroshare/provider/room.dart';
-import 'package:retroshare/services/chat.dart';
 import 'package:retroshare/ui/room/messages_tab.dart';
 import 'package:retroshare/ui/room/room_friends_tab.dart';
 import 'package:retroshare/model/chat.dart';
@@ -29,6 +29,7 @@ class _RoomScreenState extends State<RoomScreen>
   @override
   void initState() {
     super.initState();
+       
     _tabController =
         new TabController(vsync: this, length: widget.isRoom ? 2 : 1);
 
@@ -38,6 +39,7 @@ class _RoomScreenState extends State<RoomScreen>
 
     WidgetsBinding.instance.addPostFrameCallback((_) async {
       widget.chat.unreadCount = 0;
+       await registerChatEvent(context);
       await Provider.of<FriendsIdentity>(context, listen: false)
           .fetchAndUpdate();
       Provider.of<RoomChatLobby>(context, listen: false)
@@ -45,7 +47,7 @@ class _RoomScreenState extends State<RoomScreen>
       if (widget.isRoom) {
         Provider.of<RoomChatLobby>(context, listen: false)
             .updateParticipants(widget.chat.chatId);
-        await getMessagesApi(widget.chat.chatId);
+        //await getMessagesApi(widget.chat.chatId);
       }
     });
   }
