+++ lib/ui/signin_screen.dart
@@ -86,13 +86,12 @@ class _SignInScreenState extends State<SignInScreen> {
       if (value['res'] == 0 || value['res'] == 1) {
         if (value['auth']) {
           final ids = Provider.of<Identities>(context, listen: false);
-          ids.fetchOwnidenities().then((value) => {
-                ids.ownIdentity != null && ids.ownIdentity.length == 0
-                    ? Navigator.pushReplacementNamed(
-                        context, '/create_identity',
-                        arguments: true)
-                    : Navigator.pushReplacementNamed(context, '/home')
-              });
+          ids.fetchOwnidenities().then((value) {
+            ids.ownIdentity != null && ids.ownIdentity.length == 0
+                ? Navigator.pushReplacementNamed(context, '/create_identity',
+                    arguments: true)
+                : Navigator.pushReplacementNamed(context, '/home');
+          });
         } else {
           _isWrongPassword();
         }
