+++ pubspec.yaml
@@ -14,7 +14,7 @@ description: Retroshare mobile client
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.2 <3.0.0"
 
 dependencies:
   flutter:
@@ -34,7 +34,7 @@ dependencies:
   qrscan: ^0.3.2
   # Toast
   oktoast: ^2.1.8
-
+  flutter_spinkit: "^4.1.2"
   image_gallery_saver: ^1.6.9
